fix: rename module evidences to receipts
