--- src/modules/evidence/pages/EvidenceScreen/EvidenceScreen.tsx
@@ -1,1 +1,0 @@
-export const EvidenceScreen = () => <></>;
