--- src/modules/evidence/pages/EvidenceScreen/index.ts
@@ -1,1 +1,0 @@
-export * from './EvidenceScreen';
