+++ src/modules/receipts/index.ts
