+++ src/modules/receipts/pages/ReceiptsScreen/ReceiptsScreen.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const ReceiptsScreen = () => <></>;
