+++ src/modules/receipts/pages/ReceiptsScreen/index.ts
@@ -1,0 +1,1 @@
+export * from './ReceiptsScreen';
